Handle extra user menu and sidebar special pages.
This was hardcoded before, but now it is specified in configuration.
This removes a dependency of the WikimediaAPIPortal skin on the
WikimediaApiPortalOAuth extension and does not introduce a dependency
in the other direction.

For both the user menu and the sidebar, the extra items refer to
special pages. The names of these pages are localized before they
are presented to the user. To configure the wiki to resemble the
old hardcoded behavior, add the following to your LocalSettings.php:

$wgWMAPIPExtraUserMenuSpecialPages = [ 'AppManagement' ];
$wgWMAPIPSidebarSpecialPages = [
  'AppManagement',
  'Notifications',
  'Preferences'
];

Bug: T264246
Bug: T264764
Change-Id: Idcb96c29bc5354ae89d11894dfd904659bf2e9f9