Space between icons in header: space is different between search, notification and user icons when screen is narrower than 992px
Margin-right of the search icon was replaced to another file which works when a screen is wider than 992px. Because if a screen is narrower than 992px, a page will continue to use margin-right:0 for search icon. It is wrong, since there is a style which should be the same for all icons of the header when a screen is < 992px. Padding-right of the search icon was removed, because it was unnecessary and padding caused excessive space.

Bug: T267071
Change-Id: I754e1cd47bad7cb06db875b0ae91071ba1d85b59